this pattern should avoid locking the CI
